pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-agent"
version = "0.1.0"
description = "Natural-language SQL agent powered by LangChain"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "bludragon-ai"}]

dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-anthropic>=0.3,<0.4",
    "langchain-openai>=0.3,<0.4",
    "langchain-community>=0.3,<0.4",
    "sqlalchemy>=2.0,<3",
    "streamlit>=1.40,<2",
    "python-dotenv>=1.0,<2",
    "tabulate>=0.9,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
